[feature/frontend] Add settings panel section for creating, updating, and deleting domain limits (#4567)
Does what it says in the title! Also a bit of refactoring along the way to clearly separate domain permissions and domain limits, which are two separate things.

Since I'm doing this PR *before* the backend code for handling the domain limits is actually in place, there's a warning on the domain limits page that the side effects aren't implemented yet. This can be removed in later PRs that add the backend logic :) Alternatively, all the backend side effects can be implemented first, and then this PR can be updated and squerged, it doesn't matter too much!

Reviewed-on: https://codeberg.org/superseriousbusiness/gotosocial/pulls/4567
Reviewed-by: kim <gruf@noreply.codeberg.org>
Co-authored-by: tobi <tobi.smethurst@protonmail.com>
Co-committed-by: tobi <tobi.smethurst@protonmail.com>